delete button is added with password

+++ apps/server/src/modules/guns/guns.controller.ts
@@ -1,4 +1,5 @@
-import { Controller, Get, Post, Body, Param } from '@nestjs/common';
+// import { Controller, Get, Post, Body, Param } from '@nestjs/common';
+import { Controller, Get, Post, Body, Param, Headers, UnauthorizedException } from '@nestjs/common';
 import { GunsService } from './guns.service';
 import { Delete } from '@nestjs/common';
 @Controller('guns')
@@ -19,12 +20,17 @@ export class GunsController {
   findOne(@Param('id') id: string) {
     return this.gunsService.findOne(id);
   }
-  @Delete()
-  deleteAll() {
-    return this.gunsService.deleteAll();
-  }
+  // Make sure you import Headers and UnauthorizedException at the top!
+  // import { Controller, Get, Post, Body, Param, Delete, Headers, UnauthorizedException } from '@nestjs/common';
+
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  remove(@Param('id') id: string, @Headers('admin-secret') secret: string) {
+    // 1. Check the password
+    if (secret !== 'MY_SUPER_SECRET_CODE_123') {
+      throw new UnauthorizedException('Wrong password!');
+    }
+
+    // 2. If password is correct, delete it
     return this.gunsService.delete(id);
   }
 }
